Rework code - initial commit
