--- qm10/fock.py
@@ -1,12 +1,0 @@
-"""
-Functions that are used in carrying out the
-Hartree-Fock self-consistent field procedure
-"""
-
-def two_elec_energy(params):
-    """
-    This function takes molecular parameters as
-    arguments
-    """
-    return 1
-
